requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
pydantic
pydantic-settings
pydantic-core
python-jose
python-multipart
python-dotenv
requests
huggingface-hub
psycopg2-binary
APScheduler
PyYAML
httpx
openai
tflite-support
